Modify endPoints in tests. Refactor dollarEVPerTourney. Add rakebackPct in controller

+++ src/main/java/team/firestorm/repository/ModelRepository.java
@@ -36,7 +36,7 @@ public class ModelRepository {
     private Mesh mesh;
 
     private double desiredProfit;
-    private int expChipsT;
+    private double expChipsT;
     private double expEVT;
     private int tables;
     private double rakebackPercent;
